moved gnss simulation and gnss hardware into their own files
