+++ tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 from rosys.testing import forward, helpers
 
 from field_friend.automations.field_provider import Field
-from field_friend.navigation.gnss import GnssSimulation
+from field_friend.navigation.gnss_simulation import GnssSimulation
 from field_friend.system import System
 
 ROBOT_GEO_START_POSITION = [51.983159, 7.434212]
